fix load test
